Project_B. Bugfix draw result. New Menu, 1 View.

+++ Code/Project_B/Code/Data/OddParsed.cs
@@ -3,8 +3,8 @@
 namespace Project_B.Code.Data {
     public class OddParsed {
         public BetOddType Type { get; set; }
-        public decimal? AdvancedParam { get; set; }
-        public decimal Factor { get; set; }
+        public float? AdvancedParam { get; set; }
+        public float Factor { get; set; }
         public string OddStringConfirmation { get; set; }
     }
 }
